+++ app/Filament/Resources/NotificationTemplates/Pages/EditNotificationTemplate.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Filament\Resources\NotificationTemplates\Pages;
+
+use App\Filament\Resources\NotificationTemplates\NotificationTemplateResource;
+use Filament\Actions\DeleteAction;
+use Filament\Resources\Pages\EditRecord;
+
+class EditNotificationTemplate extends EditRecord
+{
+    protected static string $resource = NotificationTemplateResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            DeleteAction::make(),
+        ];
+    }
+}